rlc-cloud-repos-0.1.0/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rlc-cloud-repos"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.9"
description = "A universal cloud-init querying and repository configuration tool for Rocky Linux CIQ Derivatives"
authors = [
  { name = "Joel Hanger", email = "jhanger@ciq.com" }
]
dependencies = [
  "flake8",
  "flake8-pyproject",
  "black",
  "isort",
  "PyYAML",
  "ruamel.yaml"
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
include = ["rlc_cloud_repos"]
where = ["src"]

[tool.setuptools.package-data]
"rlc_cloud_repos" = ["data/mirrors.yaml"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
